Rename Post model to Entry and generalize it
